requirements.txt
py-trello==0.19.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.2
